# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export interface WizardSessionInput {
  projectDescription: string;
}

export interface WizardSession {
  id: number;
  title: string;
  createdAt: string;
}

export interface WizardSessionWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}